fix(lexer): prevent panic on overflowing Datadog event title/text lengths
lexEventBody guarded against declared title/text lengths exceeding the
remaining datagram bytes with a uint32 sum:

    if l.len-l.pos < l.eventTitleLen+1+l.eventTextLen {

Both lengths are attacker-controlled and each may be up to math.MaxUint32
(the per-field check in lexUint32 only rejects values strictly greater
than MaxUint32). When eventTitleLen+1+eventTextLen overflows uint32 it
wraps to a small value, so the "not enough data" guard is bypassed and
the subsequent slice expressions (e.g. input[l.pos : l.pos+eventTextLen])
panic with "slice bounds out of range".

Because the datagram parser runs in a goroutine with no recover(), a
single crafted UDP packet such as `_e{0,4294967295}:|foo` crashes the
whole gostatsd process (remote denial of service).

Perform the length arithmetic and comparison in uint64 so the guard
cannot overflow; over-long declared lengths now correctly return
errNotEnoughData, matching the guard's intent. Adds regression cases to
TestInvalidEventsLexer.

+++ internal/lexer/lexer.go
@@ -203,7 +203,12 @@ func lexDatadogSpecial(l *Lexer) stateFn {
 }
 
 func lexEventBody(l *Lexer) stateFn {
-	if l.len-l.pos < l.eventTitleLen+1+l.eventTextLen {
+	// The title and text lengths are attacker-controlled and each may be as large
+	// as math.MaxUint32. Computing the required length in uint32 can overflow and
+	// wrap to a small value, bypassing this bounds check and causing the slice
+	// operations below to panic (slice bounds out of range). Do the arithmetic and
+	// comparison in uint64 so the check cannot be bypassed.
+	if uint64(l.len-l.pos) < uint64(l.eventTitleLen)+1+uint64(l.eventTextLen) {
 		l.err = errNotEnoughData
 		return nil
 	}
